Moved function to one file so can have redundant
