+++ commands/music/save.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const constants = require('../../constants/constants')
 const fileIOUtils = require('../../utils/fileIOUtils')
+const playlistUtils = require('./musicutils/playlistutils')
 
 function saveSong(message){
     message.author.send(`You saved the track ${queue.current.title} | ${queue.current.author} from the server ${message.guild.name} ✅`).then(() => {
@@ -10,42 +11,6 @@ function saveSong(message){
     });
 
     return;
-}
-
-function savePlaylist(message, args, queue){
-    
-    var playlistName = args.join(" ").toLowerCase();
-     console.log("[Save] Saving playlist "+ playlistName);
-     
-     //Add every song URL in queue to be saved
-     var songURLs = []
-     songURLs.push(queue.current.url);
-
-     for(var i = 0;i<queue.tracks.length;i++){
-         songURLs.push(queue.tracks[i].url);
-     }
-
-     var savedPlaylist = {
-         author : message.author.username,
-         tracks : songURLs
-     }
-
-     var playslistFilePath = constants.PLAYLIST_PATH + playlistName + ".json";
-     var jsonData = JSON.stringify(savedPlaylist);
-
-     //Writing playlist to be saved
-     console.log("[Save] Writing Playlist " + playlistName);
-     fileIOUtils.writeFile(playslistFilePath,jsonData, function(err) {
-         if(err){
-             if(err = "Already Exist"){
-                 message.channel.send('Playlist already exists');
-             }
-             message.channel.send('Error saving playlist '+ playlistName);
-             return;
-         }
-
-         message.channel.send('Playlist ' + playlistName + " has been saved");
-     });
 }
 
 module.exports = {
@@ -68,9 +33,11 @@ module.exports = {
                 if(args.length < 2){
                     message.channel.send('You need to provide playlist name command : /save playlist <playlist name>');
                 }else{
-                    savePlaylist(message, args.slice(1), queue);
+                    playlistUtils.savePlaylist(message, args.slice(1), queue);
                 }
                 break;
+            default :
+                message.channel.send(`No such option for this command`);
         }
     },
 };
